Back-end's Authentication, Authorization, Registration and comfirmation are done. Only logout functionality is pending and it can be done later.
