+++ src/main/java/com/app/food/team/foodapp/config/AppConfiguration.java
@@ -11,43 +11,4 @@ import java.util.Arrays;
 
 @Configuration
 public class AppConfiguration {
-
-
-    // CORS configuration to avoid issues with the front end calls from a different address
-    // can this be done from annotation on the controller class?
-    @Bean
-    public CorsConfigurationSource corsConfigurationSource(){
-        UrlBasedCorsConfigurationSource urlBasedCorsConfigurationSource = new UrlBasedCorsConfigurationSource();
-        CorsConfiguration corsConfiguration = new CorsConfiguration();
-        corsConfiguration.setAllowCredentials(true);
-        corsConfiguration.setAllowedOrigins(Arrays.asList("http://localhost:3000","http://localhost:4200")); // React or Angular applications
-        corsConfiguration.setAllowedHeaders(Arrays.asList(
-                "Origin",
-                "Access-Control-Allow-Origin",
-                "Content-Type",
-                "Accept",
-                "Jwt-Token",
-                "Authorization",
-                "Origin, Accept",
-                "X-Requested-With",
-                "Access-Control-Request-Method",
-                "Access-Control-Request-Headers"
-        ));
-        corsConfiguration.setExposedHeaders(Arrays.asList(
-                "Origin",
-                "Access-Control-Allow-Origin",
-                "Content-Type",
-                "Accept",
-                "Jwt-Token",
-                "Authorization",
-                "Access-Control-Allow-Origin",
-                "Access-Control-Allow-Credentials",
-                "Filename"
-        ));
-        corsConfiguration.setAllowedMethods(Arrays.asList(
-                "GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS"
-        ));
-        urlBasedCorsConfigurationSource.registerCorsConfiguration("/**", corsConfiguration);
-        return urlBasedCorsConfigurationSource;
-    }
 }
