+++ src/main/java/com/app/food/team/foodapp/config/PropertiesConfiguration.java
@@ -5,6 +5,9 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 import java.security.interfaces.RSAPrivateKey;
 import java.security.interfaces.RSAPublicKey;
 
-@ConfigurationProperties(prefix = "rsa")
-public record RsaKeyPropertiesConfiguration(RSAPublicKey publicKey, RSAPrivateKey privateKey) {
+@ConfigurationProperties(prefix = "app")
+public record PropertiesConfiguration(
+        String requestMapping,
+        Integer validityTime
+    ) {
 }
