+++ src/main/java/com/app/food/team/foodapp/constants/AppConstants.java
@@ -1,0 +1,9 @@
+package com.app.food.team.foodapp.constants;
+
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+
+
+public class AppConstants {
+    private AppConstants(){}
+    public static final String BASE_URL = ServletUriComponentsBuilder.fromCurrentContextPath().build().toUriString();
+}
