+++ src/main/java/com/app/food/team/foodapp/service/EmailService.java
@@ -14,16 +14,13 @@ import org.springframework.stereotype.Service;
 @Service
 @AllArgsConstructor
 public class EmailService implements EmailServiceInterface {
-
     private final JavaMailSender mailSender;
-
     @Override
     @Async
     public void send(String to, String email) {
         try {
             MimeMessage mimeMessage = mailSender.createMimeMessage();
-            MimeMessageHelper helper =
-                    new MimeMessageHelper(mimeMessage, "utf-8");
+            MimeMessageHelper helper = new MimeMessageHelper(mimeMessage, "utf-8");
             helper.setText(email, true);
             helper.setTo(to);
             helper.setSubject("Confirm your email");
