+++ src/main/resources/application.properties
@@ -1,6 +1,3 @@
+app.request-mapping = /api/v1/
+app.validity-time = 72 
 
-request.mapping = /api/v1/
-spring.security.filter.order=10
-
-#rsa.private-key = classpath:certs/private.pem
-#rsa.public-key = classpath:certs/public.pem
